requirements.txt
asgiref==3.4.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
Babel==2.9.1
billiard==3.6.4.0
bunch==1.0.1
celery==5.0.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
chardet==4.0.0
CodeConvert==3.0.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.8
django-autoslug==1.9.8
django-bulk-update-or-create==0.3.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-constance==2.8.0
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-extra-fields==3.0.2
django-extensions==3.1.3
django-filter==2.4.0
django-import-export==2.6.1
django-jsonfield==1.4.1
django-models-ext==1.1.10
django-multiselectfield==0.1.12
django-picklefield==3.0.1
django-redis==5.0.0
django-restql==0.15.1
django-sequences==2.6
django-six==1.0.4
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg2==1.19.4
furl==2.1.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.1
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
jsonfield==3.1.0
kombu==5.1.0
MarkupSafe==2.0.1
orderedmultidict==1.0.1
packaging==21.0
Pillow==8.4.0
psycopg2-binary==2.9.1
PyMySQL==1.0.2
pycparser==2.20
PyGreSQL==5.2.4
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyPEG2==2.15.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.2
TimeConvert==1.6.0
Twisted==21.7.0
txaio==21.2.1
tzlocal==4.0.1
urllib3==1.26.7
vine==5.0.0
whitenoise==5.3.0
xlwt==1.3.0
zipp==3.6.0
zope.interface==5.4.0